maven: Refactor export integration tests into separate test projects
