+++ maven/bnd-export-maven-plugin/src/it/export-bundles-only/invoker.properties
