+++ maven/bnd-export-maven-plugin/src/it/export-from-dependencies/test.bndrun
