+++ maven/bnd-export-maven-plugin/src/it/export-from-inputbundles/bundles/org.apache.felix.eventadmin-1.4.8.jar
