+++ maven/bnd-export-maven-plugin/src/it/export-from-inputbundles/bundles/org.apache.felix.framework-5.4.0.jar
