+++ maven/bnd-export-maven-plugin/src/it/export-from-inputbundles/pom.xml
@@ -1,0 +1,40 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>biz.aQute.bnd-test</groupId>
+	<artifactId>export-from-inputbundles</artifactId>
+	<version>0.0.1</version>
+	<packaging>pom</packaging>
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>biz.aQute.bnd</groupId>
+				<artifactId>bnd-export-maven-plugin</artifactId>
+				<version>${bnd.version}</version>
+				<configuration>
+					<exporter>bnd.runbundles</exporter>
+					<bundles>
+						<bundle>${project.basedir}/bundles/org.apache.felix.eventadmin-1.4.8.jar</bundle>
+						<bundle>${project.basedir}/bundles/org.apache.felix.framework-5.4.0.jar</bundle>
+					</bundles>
+					<useMavenDependencies>false</useMavenDependencies>
+					<bndruns>
+						<bndrun>test.bndrun</bndrun>
+					</bndruns>
+					<failOnChanges>false</failOnChanges>
+					<resolve>true</resolve>
+				</configuration>
+				<executions>
+					<execution>
+						<goals>
+							<goal>export</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+</project>
