+++ maven/bnd-export-maven-plugin/src/it/export-from-inputbundles/postbuild.groovy
@@ -1,0 +1,12 @@
+//
+// The export-from-inputbundles case
+//
+
+// Check the bundle exist!
+File bundle = new File(basedir, "target/export/test/test.jar")
+assert !bundle.exists()
+
+File targetDir = new File(basedir, "target/export/test")
+assert 1 == targetDir.listFiles(new FileFilter() { boolean accept(File file) {return file.isFile();}}).length
+bundle = new File(basedir, "target/export/test/org.apache.felix.eventadmin-1.4.8.jar")
+assert bundle.exists()
