+++ maven/bnd-export-maven-plugin/src/it/export-from-inputbundles/test.bndrun
