+++ maven/bnd-export-maven-plugin/src/it/export/pom.xml
@@ -3,11 +3,6 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<parent>
-		<groupId>biz.aQute.bnd-test</groupId>
-		<artifactId>test</artifactId>
-		<version>0.0.1</version>
-	</parent>
 
 	<groupId>biz.aQute.bnd-test</groupId>
 	<artifactId>export</artifactId>
@@ -19,6 +14,21 @@
 			<plugin>
 				<groupId>biz.aQute.bnd</groupId>
 				<artifactId>bnd-export-maven-plugin</artifactId>
+				<version>${bnd.version}</version>
+				<configuration>
+					<bndruns>
+						<bndrun>test.bndrun</bndrun>
+					</bndruns>
+					<failOnChanges>false</failOnChanges>
+					<resolve>true</resolve>
+				</configuration>
+				<executions>
+					<execution>
+						<goals>
+							<goal>export</goal>
+						</goals>
+					</execution>
+				</executions>
 			</plugin>
 		</plugins>
 	</build>
