+++ maven/bnd-export-maven-plugin/src/it/export/test.bndrun
@@ -1,4 +1,4 @@
--standalone: ${.}/../index/index.xml.gz
+-standalone: ${projectsDirectory}/index/index.xml.gz
 -runfw: org.apache.felix.framework;version='[5.4.0,5.4.0]'
 -runee: JavaSE-1.8
 -runrequires: osgi.identity;filter:='(osgi.identity=org.apache.felix.eventadmin)'
