+++ maven/bnd-export-maven-plugin/src/it/index/index.xml
