+++ maven/bnd-export-maven-plugin/src/it/index/org/apache/felix/org.apache.felix.framework/5.4.0/org.apache.felix.framework-5.4.0.jar
