--- maven/bnd-export-maven-plugin/src/it/test/export-bundles-only/pom.xml
@@ -1,28 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>biz.aQute.bnd-test</groupId>
-		<artifactId>test</artifactId>
-		<version>0.0.1</version>
-	</parent>
-
-	<groupId>biz.aQute.bnd-test</groupId>
-	<artifactId>export-bundles-only</artifactId>
-	<version>0.0.1</version>
-	<packaging>pom</packaging>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>biz.aQute.bnd</groupId>
-				<artifactId>bnd-export-maven-plugin</artifactId>
-				<configuration>
-					<exporter>bnd.runbundles</exporter>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
