--- maven/bnd-export-maven-plugin/src/it/test/pom.xml
@@ -1,42 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>biz.aQute.bnd-test</groupId>
-	<artifactId>test</artifactId>
-	<version>0.0.1</version>
-	<packaging>pom</packaging>
-
-	<modules>
-		<module>export</module>
-		<module>export-bundles-only</module>
-		<module>export-from-dependencies</module>
-		<module>export-from-inputbundles</module>
-	</modules>
-
-	<build>
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>biz.aQute.bnd</groupId>
-					<artifactId>bnd-export-maven-plugin</artifactId>
-					<version>${bnd.version}</version>
-					<configuration>
-						<bndruns>
-							<bndrun>test.bndrun</bndrun>
-						</bndruns>
-						<failOnChanges>false</failOnChanges>
-						<resolve>true</resolve>
-					</configuration>
-					<executions>
-						<execution>
-							<goals>
-								<goal>export</goal>
-							</goals>
-						</execution>
-					</executions>
-				</plugin>
-			</plugins>
-		</pluginManagement>
-	</build>
-</project>
